english/requirements.txt
amqp==5.1.1
asgiref==3.7.1
billiard==4.1.0
celery==5.3.1
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coverage==5.5
Django==3.2.20
django-bootstrap-datepicker-plus==5.0.3
django-bootstrap5==22.2
django-celery-results==2.5.1
django-debug-toolbar==3.2.4
django-redis==5.3.0
flake8==5.0.4
flake8-broken-line==0.6.0
flake8-isort==5.0.0
flake8-plugin-utils==1.3.2
flake8-quotes==3.3.1
flake8-return==1.2.0
gunicorn==20.1.0
importlib-metadata==4.2.0
isort==5.10.1
kombu==5.3.1
mccabe==0.7.0
pep8-naming==0.13.2
prompt-toolkit==3.0.39
psycopg2-binary==2.9.6
pycodestyle==2.9.1
pydantic==1.10.8
pyflakes==2.5.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.5
readme-coverage-badger==0.1.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.4.0
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.6
xlwt==1.3.0
zipp==3.15.0
# uWSGI